2-Identities/hello-client/requirements.txt
msal>=1.28.0
httpx>=0.27.0
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
rich>=13.7.0
